# swagger/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Kalidou GUISSE",
        "description": "Documentation de mon API Laravel pour la gestion des clients et comptes",
        "version": "0.1"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "Serveur local de développement"
        }
    ],
    "paths": {
...
        "/clients": {
            "get": {
                "tags": [
                    "Clients"
                ],
                "summary": "Lister tous les clients",
                "description": "Retourne la liste de tous les clients.",
                "operationId": "af1f3794367a5e843bf15e3a9033d911",
                "responses": {
                    "200": {
                        "description": "Liste des clients",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Client"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnauthorizedErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur serveur",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Clients"
                ],
                "summary": "Créer un nouveau client",
                "description": "Crée un nouveau client avec les informations fournies.",
                "operationId": "15981a5da120d8b4ea1b91b05a35622b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ClientCreateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Client créé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ClientResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erreurs de validation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur serveur",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/clients/{id}": {
            "get": {
                "tags": [
                    "Clients"
                ],
                "summary": "Afficher un client spécifique",
                "description": "Retourne les détails d'un client par son ID.",
                "operationId": "2965e9737f628d15ea88209095622c24",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du client",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détails du client",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Client"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Client non trouvé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NotFoundErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur serveur",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Clients"
                ],
                "summary": "Mettre à jour un client",
                "description": "Met à jour les informations d'un client existant.",
                "operationId": "162fed3ea25bb60bf0b156c7b21313e6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du client",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ClientUpdateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Client mis à jour avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ClientResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Client non trouvé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NotFoundErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erreurs de validation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur serveur",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Clients"
                ],
                "summary": "Supprimer un client",
                "description": "Supprime un client par son ID.",
                "operationId": "20de5e8bdef993807384fd7040f9ae92",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du client",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Client supprimé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Client supprimé avec succès !"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Client non trouvé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NotFoundErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur serveur",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Client": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid",
                        "example": "c7917097-4734-4dfd-a462-b6c8f8d4aea3"
                    },
                    "nom": {
                        "type": "string",
                        "example": "Kilback"
                    },
                    "prenom": {
                        "type": "string",
                        "example": "Laury"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "brigitte.kuhn@example.net"
                    },
                    "telephone": {
                        "type": "string",
                        "example": "+1-201-736-0670"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-10-22T11:37:37.000000Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-10-22T11:37:37.000000Z"
                    }
                },
                "type": "object"
            },
...
            "ClientResponse": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Client créé avec succès !"
                    },
                    "data": {
                        "$ref": "#/components/schemas/Client"
                    }
                },
                "type": "object"
            },
            "ClientCreateRequest": {
                "required": [
                    "nom",
                    "prenom",
                    "email",
                    "telephone"
                ],
                "properties": {
                    "nom": {
                        "type": "string",
                        "example": "Kilback"
                    },
                    "prenom": {
                        "type": "string",
                        "example": "Laury"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "brigitte.kuhn@example.net"
                    },
                    "telephone": {
                        "type": "string",
                        "example": "+1-201-736-0670"
                    }
                },
                "type": "object"
            },
            "ClientUpdateRequest": {
                "properties": {
                    "nom": {
                        "type": "string",
                        "example": "Kilback"
                    },
                    "prenom": {
                        "type": "string",
                        "example": "Laury"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "brigitte.kuhn@example.net"
                    },
                    "telephone": {
                        "type": "string",
                        "example": "+1-201-736-0670"
                    }
                },
                "type": "object"
            },
            "ErrorResponse": {
                "properties": {
                    "status": {
                        "type": "string",
                        "example": "error"
                    },
                    "message": {
                        "type": "string",
                        "example": "Une erreur est survenue"
                    },
                    "errors": {
                        "type": "object",
                        "example": null,
                        "nullable": true
                    },
                    "code": {
                        "type": "integer",
                        "example": 500
                    }
                },
                "type": "object"
            },
            "ValidationErrorResponse": {
                "properties": {
                    "status": {
                        "type": "string",
                        "example": "error"
                    },
                    "message": {
                        "type": "string",
                        "example": "Les données fournies sont invalides"
                    },
                    "errors": {
                        "properties": {
                            "nom": {
                                "type": "array",
                                "items": {
                                    "type": "string"
                                },
                                "example": [
                                    "Le champ nom est obligatoire."
                                ]
                            },
                            "email": {
                                "type": "array",
                                "items": {
                                    "type": "string"
                                },
                                "example": [
                                    "Le champ email doit être une adresse email valide."
                                ]
                            }
                        },
                        "type": "object"
                    },
                    "code": {
                        "type": "integer",
                        "example": 422
                    }
                },
                "type": "object"
            },
            "NotFoundErrorResponse": {
                "properties": {
                    "status": {
                        "type": "string",
                        "example": "error"
                    },
                    "message": {
                        "type": "string",
                        "example": "Client non trouvé"
                    },
                    "errors": {
                        "type": "object",
                        "example": null,
                        "nullable": true
                    },
                    "code": {
                        "type": "integer",
                        "example": 404
                    }
                },
                "type": "object"
            },
            "UnauthorizedErrorResponse": {
                "properties": {
                    "status": {
                        "type": "string",
                        "example": "error"
                    },
                    "message": {
                        "type": "string",
                        "example": "Non authentifié"
                    },
                    "errors": {
                        "type": "object",
                        "example": null,
                        "nullable": true
                    },
                    "code": {
                        "type": "integer",
                        "example": 401
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}